+++ src/main/java/noodlezip/badge/constants/BadgeStrategyType.java
@@ -5,16 +5,17 @@ import lombok.Getter;
 @Getter
 public enum BadgeStrategyType {
 
-    SINGLE(false,null, 1, 0),
-    SINGLE_ACCUMULATIVE(true,1, 1, 0),
+    SINGLE(false, null, 1, 0),
+    SINGLE_ACCUMULATIVE(true, 1, 1, 0),
 
     LEVEL(false, null, 1, 0),
-    LEVEL_ACCUMULATIVE(true,1, 1, 0);
+    LEVEL_ACCUMULATIVE(true, 1, 1, 0);
 
     private final boolean isAccumulative;
     private final Integer initAccumulativeValue;
     private final int initCurrentValue;
     private final Integer initCurrentValueForNextLevel;
+
 
     BadgeStrategyType(boolean isAccumulative, Integer initAccumulativeValue,
                       int initCurrentValue,
