+++ src/main/java/noodlezip/badge/service/event/BadgeEventReader.java
@@ -4,6 +4,6 @@ import noodlezip.badge.dto.BadgeExtraOptionDto;
 
 public interface BadgeEventReader<T extends BadgeExtraOptionDto> {
 
-    void read(Long userId, T extraOption); //BadgeEventtype 넣어서 분기처리
+    void read(Long userId, T extraOption);
 
 }
