+++ src/main/java/noodlezip/badge/service/event/community/CommunityPostEvent.java
@@ -1,8 +1,9 @@
 package noodlezip.badge.service.event.community;
 
 import lombok.RequiredArgsConstructor;
+import noodlezip.badge.constants.LevelBadgeCategoryType;
 import noodlezip.badge.dto.EmptyInfoDto;
-import noodlezip.badge.service.category.AllCommunityPostCountBadge;
+import noodlezip.badge.service.process.level.LevelDirectUpdateProcessor;
 import noodlezip.badge.service.event.BadgeEventReader;
 import org.springframework.stereotype.Component;
 
@@ -10,12 +11,15 @@ import org.springframework.stereotype.Component;
 @Component
 public class CommunityPostEvent implements BadgeEventReader<EmptyInfoDto> {
 
-    private final AllCommunityPostCountBadge allCommunityPostCountBadge;
+    private final LevelDirectUpdateProcessor directUpdateProcessor;
 
     @Override
     public void read(Long userId, EmptyInfoDto extraOption) {
-        allCommunityPostCountBadge.process(userId);
+        processAllCommunityCount(userId);
     }
 
+    public void processAllCommunityCount(Long userId) {
+        directUpdateProcessor.process(userId, LevelBadgeCategoryType.ALL_COMMUNITY_POST_COUNT_BADGE);
+    }
 
 }
