+++ src/main/java/noodlezip/badge/service/event/review/RamenReviewPostEvent.java
@@ -1,55 +1,54 @@
 package noodlezip.badge.service.event.review;
 
 import lombok.RequiredArgsConstructor;
+import noodlezip.badge.constants.LevelBadgeCategoryType;
 import noodlezip.badge.constants.Region;
 import noodlezip.badge.dto.RamenReviewBadgeDto;
-import noodlezip.badge.service.category.AllCommunityPostCountBadge;
+import noodlezip.badge.service.process.level.LevelDirectUpdateProcessor;
+import noodlezip.badge.service.process.level.RamenReviewCategoryBadge;
 import noodlezip.badge.service.event.BadgeEventReader;
+import noodlezip.store.repository.MenuRepository;
+import noodlezip.store.repository.StoreRepository;
 import org.springframework.stereotype.Component;
+
+import java.util.List;
 
 @RequiredArgsConstructor
 @Component
 public class RamenReviewPostEvent implements BadgeEventReader<RamenReviewBadgeDto> {
 
-    private final AllCommunityPostCountBadge allCommunityPostCountBadge;
+    private final StoreRepository storeRepository;
+    private final MenuRepository menuRepository;
+    private final LevelDirectUpdateProcessor directUpdateProcessor;
+    private final RamenReviewCategoryBadge ramenReviewCategoryBadge;
 
     @Override
     public void read(Long userId, RamenReviewBadgeDto extraOption) {
-        int legalCode = getRegalCode(extraOption.getStoreId());
-
-        allCommunityPostCountBadge.process(userId);
+        processAllCommunityCount(userId);
+        processRamenCategory(userId, extraOption.getMenuIdList());
+        processRamenStoreRegion(userId, extraOption.getStoreId());
     }
 
-    private void processRamenStoreRegion(Long userId, int storeLegalCode) {
-        /**
-         * 1. 지역 파싱
-         * 2. user_id + badge_category_id + 지역 같은거 있는지 확인
-         * 3. 나머지 똑같이 연산
-         * current_score하고 total_value도 누적 저장해야함
-         */
 
+    public void processAllCommunityCount(Long userId) {
+        directUpdateProcessor.process(userId, LevelBadgeCategoryType.ALL_COMMUNITY_POST_COUNT_BADGE);
+    }
+
+    /// 라멘 카테고리 리스트 필요
+    private void processRamenCategory(Long userId, List<Long> menuIdList) {
+        // 메뉴 아이디 리스트 -> 라멘 카테고리 아이디 알아오기 -> 리스트 반목문으로 배지 로직에 전달
 
     }
 
-    private void processRamenCategory(Long userId, int ramenCategoryId) {
-        /**
-         * 찾아오는 로직만 다른거지
-         * 레벨업을 시키는 방식은 똑같다.
-         */
+    /// 지역 코드 필요
+    private void processRamenStoreRegion(Long userId, Long storeId) {
+        // 스토어 아이디 -> 법정코드 알아오기 int -> 지역코드 추출 -> 배지 로직에 전달
+        Region region = Region.getRegionByCode(11111);
+        if (region == null) {
+            return;
+        }
+        int sidoRegionCode = region.getCode();
+
     }
-
-    public int getRegalCode(Long storeId) {
-        return Region.getRegionByCode(11111).getCode();
-    }
-
-    /**
-     * 1. 배지 찾아오기
-     * 2. 초기 배지 생성 필요 확인하기 + 생성하기 - 초기 배지 생성 모듈화 가능
-     * 3. value 업데이트하기 - 전략에 따라서      - 전략에 따른 value 업데이트 모듈화 가능
-     * 4. 끝났는지 확인하기      - 모듈화 가능
-     * 4.
-     */
-
-
 
 }
