--- src/main/java/noodlezip/badge/service/process/level/handler/InitLevelBadgeFactory.java
@@ -1,4 +1,0 @@
-package noodlezip.badge.service.process.level.handler;
-
-public class InitLevelBadgeFactory {
-}
