+++ src/main/java/noodlezip/badge/service/process/level/handler/ValueUpdateHandler.java
@@ -5,7 +5,7 @@ import noodlezip.badge.entity.UserBadge;
 import org.springframework.stereotype.Component;
 
 @Component
-public class LevelValueUpdateHandler { //인터페이스로 놔도 될듯,
+public class ValueUpdateHandler {
 
     public boolean handle(UserBadge userBadge, BadgeStrategyType strategy) {
         if (strategy.equals(BadgeStrategyType.LEVEL)) {
@@ -16,7 +16,7 @@ public class LevelValueUpdateHandler { //인터페이스로 놔도 될듯,
         return false;
     }
 
-    private boolean updateLevelAccumulativeTypeBadgeValue(UserBadge userBadge) {
+    public boolean updateLevelAccumulativeTypeBadgeValue(UserBadge userBadge) {
         userBadge.updateAccumulativeValueByOne();
         if (!userBadge.isFinalBadgeAlreadyObtained()) {
             userBadge.updateCurrentValueByOne();
@@ -25,7 +25,7 @@ public class LevelValueUpdateHandler { //인터페이스로 놔도 될듯,
         return false;
     }
 
-    private boolean updateLevelTypeBadgeValue(UserBadge userBadge) {
+    public boolean updateLevelTypeBadgeValue(UserBadge userBadge) {
         if (userBadge.isFinalBadgeAlreadyObtained()) {
             return false;
         }
